switched from gaze to node-watch

+++ cli.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-var gaze = require('gaze');
+var watch = require('node-watch');
 var path = require('path');
 var hamlc = require('haml-coffee');
 var assert = require('util').assert;
@@ -13,29 +13,34 @@ if (process.argv.length === 2) {
 
 var watchDirs = process.argv.slice(2).map(f => path.resolve(f))
 
-var gazers = watchDirs.map(function (dir) {
-    return gaze('*.haml', {cwd: dir});
-});
+watch(watchDirs, onChange);
 
-gazers.forEach(function (gazer) {
-    gazer.on('all', onAll);
-});
-
-
-function onAll (event, hamlFilepath) {
-    htmlFilepath = hamlFilepath.slice(0, hamlFilepath.length - '.haml'.length) + '.html';
-    switch (event) {
-        case 'added':
-        case 'changed':
-            console.log(`compiling ${asRelative(hamlFilepath)} to ${asRelative(htmlFilepath)}`);
-            haml = fs.readFileSync(hamlFilepath, 'utf-8');
-            html = hamlc.render(haml);
-            fs.writeFileSync(htmlFilepath, html);
-            break;
-        case 'deleted':
+function onChange(filepath) {
+    if (path.extname(filepath) !== '.haml') {
+        return;
+    }
+    var hamlFilepath = filepath;
+    var htmlFilepath = hamlFilepath.slice(0, hamlFilepath.length - '.haml'.length) + '.html';
+    try {
+        haml = fs.readFileSync(hamlFilepath, 'utf-8');
+        console.log(`compiling ${asRelative(hamlFilepath)} to ${asRelative(htmlFilepath)}`);
+        html = hamlc.render(haml);
+        fs.writeFileSync(htmlFilepath, html);
+    } catch (err) {
+        if (err.code === "ENOENT") {
             console.log(`deleting ${asRelative(htmlFilepath)}`);
-            fs.unlinkSync(htmlFilepath);
-            break;
+            try {
+                fs.unlinkSync(htmlFilepath);
+            } catch (err) {
+                if (err.code === "ENOENT") {
+                    console.log(`file ${asRelative(htmlFilepath)} was already deleted.`);
+                } else {
+                    throw(err);
+                }
+            }
+        } else {
+            throw(err);
+        }
     }
 }
